Update README.md
Given a csv file with lines of order data, find duplicate product purchases within that file (comparing rows within the file for similarlity); also compare this file with another file with a list of product names to identify similar items. Note that the text may be mispelled, abbreviated, etc. So think fuzzy searching.